Refactor end screen
Moved functions displaying end_screen and wake_up screen into different file for every variant to use
