+++ src/round.py
@@ -210,7 +210,7 @@ class Round:
 
     def use_card(self, game_round, game_renderer, state, card_from_stack):
         # nie wiem czy tu ma byc card from stack
-        Player.human_use_ability(self, state, Card, game_round, game_renderer)
+        Player.human_use_ability(self, state, ard, game_round, game_renderer)
         card_from_stack.show_front = False
         game_renderer.draw_state(game_round, state, "Używasz karty")
         pygame.time.wait(1000)
@@ -399,7 +399,7 @@ class Round:
         game_round.debug(state)
         game_renderer.draw_state(game_round, state, "Zamieniono miejscami")
 
-    def bot_take_card_from_any_pile(self, state, game_round, game_renderer):
+    def bot_take_card_from_any_pile_POP(self, state, game_round, game_renderer):
         print("robot! ᕙ(  •̀ ᗜ •́  )ᕗ") #do zrobienia
